+++ test/runtests.jl
@@ -38,6 +38,8 @@ include("common.jl")
 
     include("data.jl")
 
+    include("opb.jl")
+
     include("pf.jl")
 
     include("opf.jl")