+++ CHANGELOG.md
@@ -3,6 +3,7 @@ PowerModels.jl Change Log
 
 ### Staged
 - Added support for current limits, issue #342
+- Added a optimal power balance problem specification
 - Added a network flow approximation formulation, NFAPowerModel
 - Added data simplification for piecewise linear cost functions
 - Added sim_parallel_run_time to OBBT stats